Fixed Truss, added Frame
Fixed a data type issue with the truss solving that cause the
displacements to round to an integer.
