+++ Examples/Truss 3.py
@@ -3,17 +3,18 @@ sys.path.append('..')
 import structural_classes as sc
 import cross_sections as xs
 import materials as ma
+import Truss
 
 # Define material
 A992 = ma.A992()
 
 # Define cross section
-xs1 = xs.generalSection(A=0.01)
+xs1 = xs.generalSection(A=1)
 
 # define blank structure
 # the cross section and material are defaults for the members
 # we will add to this later
-s1 = sc.Structure(cross=xs1, material=A992)
+s1 = Truss.Truss(cross=xs1, material=A992)
 
 # Add nodes to the structure
 s1.addNode(0, 0, xfix=0, yfix=0)
@@ -27,12 +28,12 @@ s1.addMember(0, 2)
 
 s1.plot()
 
-Forces = np.matrix('0; 0; 8660; 5000; 0; 0')
-displacement = s1.directStiffness(Forces)
-s1.plotDeformation(nfig=1)
+Forces = np.matrix('0; 0; 1000; 1000; 0; 0')
+disp = s1.directStiffness(Forces)
+s1.plotDeformation(nfig=1, mag=1)
 
-Forces = np.matrix('0; 0; 0; -100000; 0; 0')
+Forces = np.matrix('0; 0; 100; 100; 0; 0')
 s1.directStiffness(Forces)
-s1.plotDeformation(nfig=2)
+s1.plotDeformation(nfig=2, mag=1)
 
 s1.printMembers()
